quantum_database.py
import os
import random
import qiskit
from qiskit import Aer, execute, IBMQ
from qiskit.circuit import QuantumCircuit
from dotenv import load_dotenv
import google.generativeai as gen_ai

# Load environment variables
load_dotenv()

# Setup Google Gemini API Key
API_KEY = os.getenv('GOOGLE_API_KEY')
gen_ai.configure(api_key=API_KEY)

# Setup IBM Quantum Token (for Qiskit)
IBM_TOKEN = os.getenv('IBM_TOKEN')
IBMQ.save_account(IBM_TOKEN, overwrite=True)

# Simulate a quantum database query using Qiskit
def query_quantum_database(query):
    try:
        if query[0] == "quantum_search":
            # Use Qiskit's Aer simulator to simulate a quantum search
            backend = Aer.get_backend('statevector_simulator')
            qc = QuantumCircuit(2)
            qc.h([0, 1])
            qc.cz(0, 1)

            # Simulate without measurement for statevector
            job = execute(qc, backend)
            result = job.result().get_statevector()

            return f"Quantum Search Result: {result}"

        elif query[0] == "quantum_data":
            # Simulate fetching quantum data
            simulated_data = {
                'data1': [random.random() for _ in range(3)],
                'data2': [random.random() for _ in range(3)],
                'data3': [random.random() for _ in range(3)],
            }
            data = simulated_data.get(query[1], "No data found for this query")
            return f"Quantum Data Result for {query[1]}: {data}"

        else:
            return "Invalid query type."

    except Exception as e:
        return f"Error querying quantum database: {str(e)}"

