requirements.txt
lambeq==0.3.0
pytket==1.18.0
pytket-qiskit
pennylane
numpy
streamlit
python-dotenv
google-generativeai
